Created a Django Search app that uses the arXiv API
Use the arXiv API to search for papers. Creates a search functionality under

http://localhost:8000/search/<search-term>/<num_results>/

that queries the arXiv API and returns the desired number of results. Currently
filters 1. the arXiv ID, 2. the title, 3. the author(s), 4. the published date
and 5. the abstract. Currently renders the results using an html template
but we should pass this information back to React to display to the users.